feat: created tutor db model, tutor controller and tutor route, that then goes into server.ts
