+++ package.json
@@ -6,7 +6,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "prepare": "husky",
     "compile": "npx tsc",
-    "server": "nodemon",
+    "server": "npx tsx watch ./src/server.ts",
     "runfile": "npx nodemon --exec tsx"
   },
   "keywords": [],
