+++ src/models/getTutor.ts
@@ -6,10 +6,14 @@ const tutors = zubiDB.prepare("SELECT * FROM tutors");
 
 const allRows = tutors.all();
 
-const getTutorById = (index: number) => {
-  return allRows[index - 1];
+// const getTutorById = (index: number) => {
+//   return allRows[index - 1];
+// };
+
+const getTutorById = () => {
+  return allRows;
 };
 
-console.log(getTutorById(2));
+// console.log(getTutorById(2));
 
 export default getTutorById;
